pyproject.toml
[project]
name = "todo-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-nested-routers>=0.95.0",
    "drf-spectacular>=0.29.0",
    "pillow>=12.2.0",
    "psycopg2-binary>=2.9.12",
    "python-decouple>=3.8",
]
